Use Unix epoch time for build number
Google Camera 8.2.400 crashes when attempting to take a picture with
Night Sight on custom ROMs because the build number doesn't match the
expected format: a signed 32-bit integer.

The parsed number is used as a gate for a burst-related change. The
minimum threshold is 7168887 for RQ3A builds. The current Unix epoch
time in seconds happens to be greater than this and conveys the same
information as the old eng build number (except the username, which we
don't want there anyway), so let's use it as the incremental build
number.

NB: We change BUILD_NUMBER here instead of vendorsetup because it needs
to be unique to each build, not lunch session.

Change-Id: If2b0caade2fa874719dd971a9ea7fce37f0b2af4